Add test for signup success flow
