+++ examples/starknet-react-next/package.json
@@ -6,6 +6,7 @@
     "dev": "next dev -p 3002",
     "build": "next build",
     "test": "playwright test",
+    "test:ui": "playwright test --ui",
     "start": "next start -p 3002",
     "lint": "next lint",
     "format": "prettier --write ."
